feat!: support `aws-lc-rs` or `rust-crypto`; remove `ring`
Add `aws-lc-rs` and `rust-crypto` features, exactly one of which is
required to be enabled by `jsonwebtoken` v10. `rust-crypto` is now
part of `default` features.

Use new `jsonwebtoken::dangerous::insecure_decode` to support
jsonwebtoken v10, which introduced breaking changes to the previous
workaround to insecurely decode a JWT.

BREAKING CHANGE: `rust-crypto` and `aws-lc-rs` are mutually exclusive so
`--all-features` will fail. This is an inherent limitation of
`jsonwebtoken` v10 at this time.

BREAKING CHANGE: dependency on `ring` is eliminated and replaced by
either `aws-lc-rs` OR `rust-crypto`.

Signed-off-by: Dwayne Sykes <dwayne@sykes.pw>